changed printData signature to using size_t in heatRead example

+++ examples/heatTransfer/read/PrintData.cpp
@@ -15,8 +15,7 @@
 
 #include "PrintData.h"
 
-void printData(double *xy, uint64_t *size, uint64_t *offset, int rank,
-               int steps)
+void printData(double *xy, size_t *size, size_t *offset, int rank, int steps)
 {
     std::ofstream myfile;
     std::string filename = "data." + std::to_string(rank);
